feature-gate most schedulers
