+++ Cargo.toml
@@ -6,18 +6,20 @@ edition = "2021"
 [features]
 datapath = ["tun-tap", "libc", "socket2", "pnet", "flume"]
 bin = ["datapath", "clap", "tracing-subscriber", "htb-argparse"]
-afd = ["rand"]
-htb-argparse = ["clap"]
+afd = ["rand", "yaml-rust"]
+hdwrr = ["yaml-rust"]
+htb = ["csv", "serde"]
+htb-argparse = ["htb", "clap"]
 hwfq-audit = []
 
 [dependencies]
 quanta = "0.12"
-yaml-rust = "0.4"
 tracing = "0.1"
 color-eyre = "0.6"
 etherparse = "0.14"
-csv = "1.3"
-serde = { version = "1", features = ["derive"] }
+yaml-rust = { version = "0.4", optional = true }
+csv = { version = "1.3", optional = true }
+serde = { version = "1", features = ["derive"], optional = true }
 tun-tap = { version = "0.1", default-features=false, features = [], optional = true }
 socket2 = { version = "0.5.3", features = ["all"], optional = true }
 pnet = { version = "0.29", optional = true }
@@ -32,12 +34,16 @@ tracing-subscriber = "0.3"
 
 [[bin]]
 name = "hwfq"
-required-features = ["bin", "afd"]
+required-features = ["bin", "afd", "hdwrr"]
 
 [[bin]]
 name = "policer"
-required-features = ["bin"]
+required-features = ["htb", "bin"]
 
 [[bin]]
 name = "uds_out"
 required-features = ["bin"]
+
+[[example]]
+name = "weight_tree"
+required-features = ["hdwrr"]
